pyproject.toml
[project]
name = "prism-tooling"
version = "0.1.0"
description = "Python tooling for Prism data access gateway"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "docker>=7.0.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "protobuf>=4.25.1",
    "grpcio-tools>=1.60.0",
    "black>=23.12.1",
    "mypy>=1.8.0",
    "python-frontmatter>=1.1.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tooling"]

[dependency-groups]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.9",
]

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
